Some cleanup of the ws-security runtime code
git-svn-id: https://svn.apache.org/repos/asf/cxf/trunk@1333402 13f79535-47bb-0310-9956-ffa450edef68

+++ rt/ws/security/src/main/java/org/apache/cxf/ws/security/trust/STSClient.java
@@ -1195,7 +1195,7 @@ public class STSClient implements Configurable, InterceptorProvider {
     protected void addLifetime(XMLStreamWriter writer) throws XMLStreamException {
         Date creationTime = new Date();
         Date expirationTime = new Date();
-        expirationTime.setTime(creationTime.getTime() + (ttl * 1000L));
+        expirationTime.setTime(creationTime.getTime() + ((long)ttl * 1000L));
 
         XmlSchemaDateFormat fmt = new XmlSchemaDateFormat();
         writer.writeStartElement("wst", "Lifetime", namespace);
